requirements.txt
flask==3.0.0
gunicorn==21.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.2
ultralytics==8.1.0
pillow==10.2.0
mediapipe==0.10.9
requests==2.31.0
supervision==0.18.0
